add temporary sign-up and sign-in screens
