+++ .gitignore
@@ -18,7 +18,7 @@
 # misc
 .DS_Store
 *.pem
-.idea/*
+/.idea/
 
 # debug
 npm-debug.log*
